+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ### 0.1.5r (WIP)
-- APNG support
+- Added APNG support
+- Added option to specify output filename
+- Added option to use custom .PAL file
+- Added option to convert multiple .FRM files
+- Added option to select .PNG generator
 
 ### 0.1.3 (2018-01-04)
 - AppVeyor configuration (alexeevdv)